backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.4
python-dotenv
requests
pytesseract
Pillow
python-multipart
genai
